Add native prop to signal reversed lists
Summary: While we certainly can check the order of content container children in a scroll view to determine if a list is in reverse order, this doesn't tell us definitively whether the semantic intent of the list is reverse ordered items. Instead, we can pass the same indicator that we use to determine list cell metrics to the native component.

Differential Revision: D64759218